Minimalisir error (#235)
* k

* style: format code with black and isort (#232)

Format code with black and isort

This commit fixes the style issues introduced in 1bb7a44 according to the output
from Black and isort.

Details: https://app.deepsource.com/gh/yasirarism/MissKatyPyro/transform/ee1290fb-7c97-4a0f-9b7e-8de11488be3c/


* co ba

* style: format code with black and isort (#233)

Format code with black and isort

This commit fixes the style issues introduced in 94392c1 according to the output
from Black and isort.

Details: https://app.deepsource.com/gh/yasirarism/MissKatyPyro/transform/e7ec7f17-5ed0-42dc-9f15-9deb6c3500cb/


* tess

* style: format code with black and isort (#234)

Format code with black and isort

This commit fixes the style issues introduced in 06a6c7f according to the output
from Black and isort.

Details: https://app.deepsource.com/gh/yasirarism/MissKatyPyro/transform/8e5e9d83-441b-49e4-a3ce-7c731912825f/


* Update tes_session.py

* Update urban_dict.py

* Update web_scraper.py

---------

+++ misskaty/plugins/webss.py
@@ -44,6 +44,7 @@ async def take_ss(_, ctx: Message, strings):
             ]
         )
         await msg.delete_msg()
-        os.remove(download_file_path)
+        if os.path.exists(download_file_path):
+            os.remove(download_file_path)
     except Exception as e:
         await msg.edit_msg(strings("ss_failed_str").format(err=str(e)))
